moved get_region_slicer
